(chore) rename nvidia provider to openaiCompatible and split ssrf guard
